requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
# Django==4.0.4
Django==3.2.18
django-cors-headers==3.12.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
MarkupSafe==2.1.1
# numpy==1.22.4
numpy==1.21.6
packaging==21.3
# pandas==1.4.2
pandas==1.3.5
patsy==0.5.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
PyYAML==6.0
requests==2.27.1
# scikit-learn==1.1.1
scikit-learn==1.0.1
# scipy==1.8.1
scipy==1.7.3
six==1.16.0
sqlparse==0.4.2
statsmodels==0.13.2
threadpoolctl==3.1.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.1.0
